- Make mem-usage counters use size_t and use the %zu conversion specifier accordingly
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/trunk@6719 82007160-df01-0410-b94d-b575c5fd34c7

+++ include/ipcache.h
@@ -46,6 +46,6 @@ struct ip_entry
 
 extern struct ip_entry *ipcache_find_or_add_address(const struct irc_ssaddr *);
 extern void ipcache_remove_address(const struct irc_ssaddr *);
-extern void ipcache_get_stats(unsigned int *const, uint64_t *const);
+extern void ipcache_get_stats(unsigned int *const, size_t *const);
 extern void ipcache_init(void);
 #endif
